Safari-609.1.11.1/LayoutTests/fast/repaint/outline-with2px-auto-repaint-rect.html

<html>
<script>jsTestIsAsync = true;</script>
<script src="../../resources/js-test-pre.js"></script>
<head>
<title>This tests that outline repaint rect for focus ring is correct.</title>
<style>
div {
  position: absolute;
  top: 10px;
  height: 10px;
  width: 10px;
  outline: auto 2px;
}
</style>
</head>
<body>
<div id=foo></div>
</body>
<script>
setTimeout(
	function() { 
        if (window.internals)
	        internals.startTrackingRepaints();
 		document.getElementById("foo").style.top = "50px";
 		document.body.offsetWidth;

        if (window.internals) {
            shouldNotBe("window.internals.repaintRectsAsText().indexOf('5 7')", "-1");
            shouldNotBe("window.internals.repaintRectsAsText().indexOf('5 47')", "-1");
            internals.stopTrackingRepaints();
        }
        finishJSTest();
	}, 0);
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
