LayoutTests/fast/repaint/outline-with2px-auto-repaint-rect-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('5 7') is not -1
PASS window.internals.repaintRectsAsText().indexOf('5 47') is not -1
PASS successfullyParsed is true

TEST COMPLETE
